requirements.txt
PyQt6
tkcalendar
flask
flask-cors
flask-restful
PyJWT
